[ci skip] Users can now switch between several roles.
Q: when we switch role, the page is reloaded isn't it? => what if the current page becomes forbidden (no menu is displayed on the 403 page, so only "BACK" to the last allowed page is possible)
Q: avoid deletion of a "secondary" role (in M2M but currently not used as role by a user)?

TODO: menu entries to switch
 - WAIT for get_context() in MenuEnty ??
 - reload page
   => what if the current page becomes forbidden => menu is displayed?
 - repair Users with active role not contained in M2M 'roles'

TODO: clean
 remaining TODOs in code
TODO: i18n (config, core, persons)
TODO: changelog
 - feature
 - role.ordering => ('name',)
 - breaking
    - user validation improved